refactor: /freqaimodels to webserver section
